+++ src/main/groovy/com/deploygate/gradle/plugins/tasks/UploadAabTask.groovy
@@ -8,6 +8,7 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.Internal
+import org.gradle.api.tasks.Nested
 import org.gradle.api.tasks.TaskAction
 import org.jetbrains.annotations.NotNull
 import org.jetbrains.annotations.VisibleForTesting
@@ -27,7 +28,7 @@ abstract class UploadAabTask extends UploadArtifactTask {
                 )
     }
 
-    @Internal
+    @Nested
     final Property<AabInfo> aabInfo
 
     @Inject
@@ -47,11 +48,5 @@ abstract class UploadAabTask extends UploadArtifactTask {
         def inputParams = inputParamsProvider.get()
 
         doUpload(inputParams)
-    }
-
-    @Internal
-    @Override
-    String getDescription() {
-        return "Deploy bundled ${inputParamsProvider.get().variantName} to DeployGate"
     }
 }